LayoutTests/platform/wpe/imported/w3c/web-platform-tests/webxr/xrView_eyes.https-expected.txt


PASS XRView.eye is correct for immersive sessions - webgl
PASS XRView.eye is correct for immersive sessions - webgl2
PASS XRView.eye is correct for non-immersive sessions - webgl
PASS XRView.eye is correct for non-immersive sessions - webgl2
